refactor(educational-content-index): restructure search and storage modules
- Split search functionality into engine and filters modules
- Move storage-related code from metadata to dedicated storage module
- Update imports and module structure accordingly
