+++ packages/soroban/contracts/educational-content-index-contract/src/search/mod.rs
@@ -1,0 +1,5 @@
+pub mod engine;
+pub mod filters;
+
+// Re-export the main search function for backward compatibility
+pub use engine::search_content;
